requirements.txt
# ============================================
# Big Data Platform - Python Dependencies
# ============================================

# ==========================================
# KAFKA & Streaming
# ==========================================
kafka-python==2.0.2
confluent-kafka==2.3.0
avro-python3==1.10.2
fastavro==1.9.0

# ==========================================
# SPARK & Data Processing
# ==========================================
pyspark==3.5.0
pyarrow==14.0.1
fastparquet==2023.10.1

# ==========================================
# CLOUD & Storage
# ==========================================
# Google Cloud (BigQuery Data Warehouse)
google-cloud-storage==2.13.0
google-cloud-bigquery==3.14.1
google-auth==2.25.2

# MinIO (Data Lake S3-compatible)
boto3==1.34.6
minio==7.2.0
s3fs==2024.2.0

# ==========================================
# WEB SCRAPING
# ==========================================
# Core scraping frameworks
scrapy==2.11.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Browser automation (pour sites JavaScript)
selenium==4.16.0
playwright==1.40.0
webdriver-manager==4.0.1

# HTTP clients
requests==2.31.0
requests-html==0.10.0
httpx==0.25.2

# Anti-ban & Proxies
scrapy-rotating-proxies==0.6.2
fake-useragent==1.4.0
scrapy-user-agents==0.1.1

# ==========================================
# NLP & Text Processing
# ==========================================
# Core NLP
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Language detection
langdetect==1.0.9
pycld2==0.41

# Text preprocessing
unidecode==1.3.7
ftfy==6.1.3

# ==========================================
# CV/RESUME PARSING
# ==========================================
# PDF parsing
pdfplumber==0.10.3
PyPDF2==3.0.1
pdfminer.six==20221105

# Word documents
python-docx==1.1.0
python-pptx==0.6.23

# OCR (si besoin)
pytesseract==0.3.10

# ==========================================
# DATA MANIPULATION
# ==========================================
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1

# ==========================================
# AIRFLOW
# ==========================================
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.5.0
apache-airflow-providers-google==10.12.0
apache-airflow-providers-amazon==8.14.0

# ==========================================
# DATABASE
# ==========================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1

# ==========================================
# UTILITIES
# ==========================================
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1
jsonschema==4.20.0
validators==0.22.0

# Date & time
python-dateutil==2.8.2
pytz==2023.3.post1

# Logging & monitoring
loguru==0.7.2
structlog==23.2.0

# Performance
multiprocessing-logging==0.3.4
celery==5.3.4

# ==========================================
# DEVELOPMENT & TESTING
# ==========================================
jupyter==1.0.0
jupyterlab==4.0.9
ipython==8.18.1
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# ==========================================
# SUPERSET (BI)
# ==========================================
# Note: Superset est dans son propre conteneur Docker
# Ces dépendances sont pour l'intégration si nécessaire
apache-superset==3.0.3

# ==========================================
# VERSIONING
# ==========================================
# Note: Fichier généré le 2024-11-24
# Python 3.11+ recommandé
# Pour mettre à jour: pip install --upgrade -r requirements.txt
